refactor: align Metrics service with TM1py flat-module conventions
Collapse the TM1py/Metrics/ sub-package (vocabulary, mdx, odata_filter,
shapers) into TM1py/Services/MetricService.py as module-level helpers, matching
the single-flat-file pattern every other service follows. Consolidate the five
Metric*_test.py files into one Tests/MetricService_test.py per the
one-test-file-per-service convention. Fold usage examples into the service
docstring so they render in the auto-generated mkdocs API Reference. Remove ADR/
PRD references from docstrings and comments.

No behavior change: 44 unit tests pass; black and ruff clean.
